+++ applications/aws_dashboard/pages/endpoints/page.py
@@ -8,8 +8,8 @@ from sageworks.views.artifacts_web_view import ArtifactsWebView
 from sageworks.web_components import table
 
 # Local Imports
-from pages.layout.endpoints_layout import endpoints_layout
-import pages.callbacks.endpoints_callbacks as callbacks
+from .layout import endpoints_layout
+from . import callbacks
 
 register_page(__name__, path="/endpoints")
